api/requirements.txt
Django==4.1
djangorestframework==3.14

uuid
pillow

# Database
dj_database_url==2.1.0
psycopg2-binary==2.9.9
